Cohen Sutherland Line Clipping Algorithm in C
This is one of the oldest and most popular line clipping algorithm. To speed up the process this algorithm performs initial tests that reduce number of intersections that must be calculated. It does so by using a 4 bit code called as region code or outcodes. These codes identify location of the end point of line.Each bit position indicates a direction, starting from the rightmost position of each bit indicates left, right, bottom, top respectively.